move isLocalNode to hook
* pass isLocalNode to FormatReceivedBy
