+++ graylog2-web-interface/src/views/components/messagelist/FormatReceivedBy.tsx
@@ -15,7 +15,6 @@
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
 import * as React from 'react';
-import { useState, useEffect } from 'react';
 import * as Immutable from 'immutable';
 
 import { Spinner } from 'components/common';
@@ -32,20 +31,9 @@ const _inputName = (inputs: Inputs, inputId: string) => {
   return input ? <span style={{ wordBreak: 'break-word' }}>{input.title}</span> : 'deleted input';
 };
 
-const FormatReceivedBy = ({ inputs, sourceInputId, sourceNodeId }: { inputs: Inputs, sourceNodeId: string, sourceInputId: string }) => {
-  const [isLocalNode, setIsLocalNode] = useState<boolean | undefined>();
-
+const FormatReceivedBy = ({ isLocalNode, inputs, sourceInputId, sourceNodeId }: {isLocalNode: boolean, inputs: Inputs, sourceNodeId: string, sourceInputId: string }) => {
   const forwarderPlugin = usePluginEntities('forwarder');
   const ForwarderReceivedBy = forwarderPlugin?.[0]?.ForwarderReceivedBy;
-  const _isLocalNode = forwarderPlugin?.[0]?.isLocalNode;
-
-  useEffect(() => {
-    if (sourceNodeId && _isLocalNode) {
-      _isLocalNode(sourceNodeId).then(setIsLocalNode, () => setIsLocalNode(true));
-    } else {
-      setIsLocalNode(true);
-    }
-  }, [sourceNodeId, _isLocalNode]);
 
   if (!sourceNodeId) {
     return null;
